requirements.txt
# Core dependencies
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.3.0
langgraph==0.2.45
langsmith==0.1.147

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
pgvector==0.3.6
asyncpg==0.30.0

# API
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
pydantic-settings==2.6.1
python-multipart==0.0.17
websockets==13.1

# Data processing
pandas==2.2.3
numpy==2.1.3
pyarrow==18.1.0

# Visualization
plotly==5.24.1
kaleido==0.2.1

# Frontend
streamlit==1.40.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
click==8.1.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

# Development
black==24.10.0
ruff==0.8.0
mypy==1.13.0
pre-commit==4.0.1

# Monitoring (optional)
langfuse==2.54.2
structlog==24.1.0
